+++ actix-client-ip-cloudflare/CHANGELOG.md
@@ -2,6 +2,12 @@
 
 ## Unreleased
 
+- Add `TrustedIps::new()` constructor.
+- Add `TrustedIps::add_ip_range()` method.
+- Add `TrustedIps::{add_loopback_ips, add_private_ips}()` methods.
+- Implement `Default` for `TrustedIps`.
+- Deprecate `TrustedIps::with_ip_range()` method.
+
 ## 0.1.0
 
 - Initial release.